Refactor undo behavior into controller and reintroduce controller as middleman
